Don't set timezone explicitly to UTC
The maven master sets the timezone only for the surefire and failsafe
plugins. When a user adds other plugins like cargo he'd need to remember
adding this timezone as well to avoid hard-to-diagnose errors.

It is simpler to use the hosts timezone here so that all plugins have a
consistent setting without user configuration.

Change-Id: Ibf4f924ce56a7cdd162ac644a0a28818ae44c730