Avoid copying content of src/main/webapp to output folder twice
- Removed webapp_folder profile 
- Building a project using <packaging>war</packaging> copies the content
of src/main/webapp to the root output folder (root of WAR file) as
default. 
- Using the webapp_folder profile in addition leads to a duplication of
the content of src/main/webapp, the webapp_folder profiles copies the
content of src/main/webapp to the default output folder (e.g.
target/classes or WEB-INF/classes), therefore the content of
src/main/webapp is available twice in the final WAR file:
 1. in root of WAR (OK)
 2. in \WEB-INF\classes  (NOK)

204640

Change-Id: I6358b3a9cef5bf62cedc79c3af2f34c70daf3305
Signed-off-by: Paolo Bazzi <paolo.bazzi@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/89874
Reviewed-by: Patrick B??nziger <patrick.baenziger@bsi-software.com>
Tested-by: Patrick B??nziger <patrick.baenziger@bsi-software.com>