Do not delete node_modules in clean phase
It is not possible to safely clean the node_modules folder because the
maven-clean-plugin cannot handle symlinks correctly and deletes all
linked files instead of the symlink only. <followSymlinks> seems not to
work correctly. Therefore do not remove the node_modules folder and rely
on pnpm to update the folder correctly. This also improves build
performance because the node_modules folder might be reused.