Fix eclipse endless maven update
The maven-jar-plugin will generate files like the MANIFEST.MF. As this
file contains the build timestamp it changes its content on every maven
update. m2e will listen on these updates and trigger a new maven update
of the corresponding module and therefore ends in an endless loop.
Disable the maven-jar-plugin to break this endless loop.

374952