Prevent info message in resource filtering.
If the property is not set, the following message might appear:
The encoding used to copy filtered properties files have not been set.
This means that the same encoding will be used to copy filtered
properties files as when copying other filtered resources. This might
not be what you want!

This is actually what we want and therefore explicitly set the encoding
for properties to prevent this confusing message. It has no effect on
the actual build